Start groundwork to use non-default NIC
Not really Parallels specific, i.e. wanted to do this on Hyper-V, but since it was faster to get here than with Hyper-V, will start getting it not use the default NIC. Was hoping to not need it, but seems DHCP on this version of Parallels behave funny, i.e. I have duplicate IP assignments